+++ packages/react/src/utils/changes.ts
@@ -1,4 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
+import { EdgeLookup, NodeLookup } from '@xyflow/system';
 import type { Node, Edge, EdgeChange, NodeChange, NodeSelectionChange, EdgeSelectionChange } from '../types';
 
 export function handleParentExpand(updatedElements: any[], updateItem: any) {
@@ -53,110 +54,118 @@ export function handleParentExpand(updatedElements: any[], updateItem: any) {
 // When you drag a node for example, React Flow will send a position change update.
 // This function then applies the changes and returns the updated elements.
 function applyChanges(changes: any[], elements: any[]): any[] {
-  // we need this hack to handle the setNodes and setEdges function of the useReactFlow hook for controlled flows
-  if (changes.some((c) => c.type === 'reset')) {
-    return changes.filter((c) => c.type === 'reset').map((c) => c.item);
-  }
-
-  let remainingChanges = [];
   const updatedElements: any[] = [];
+  // By storing a map of changes for each element, we can a quick lookup as we
+  // iterate over the elements array!
+  const changesMap = new Map<any, any[]>();
 
   for (const change of changes) {
     if (change.type === 'add') {
       updatedElements.push(change.item);
+      continue;
+    } else if (change.type === 'remove' || change.type === 'replace') {
+      // For a 'remove' change we can safely ignore any other changes queued for
+      // the same element, it's going to be removed anyway!
+      changesMap.set(change.id, [change]);
     } else {
-      remainingChanges.push(change);
+      const elementChanges = changesMap.get(change.id);
+
+      if (elementChanges) {
+        // If we have some changes queued already, we can do a mutable update of
+        // that array and save ourselves some copying.
+        elementChanges.push(change);
+      } else {
+        changesMap.set(change.id, [change]);
+      }
     }
   }
 
-  for (const item of elements) {
-    const nextChanges: any[] = [];
-    const _remainingChanges: any[] = [];
+  for (const element of elements) {
+    const changes = changesMap.get(element.id);
 
-    for (const change of remainingChanges) {
-      if (change.id === item.id) {
-        nextChanges.push(change);
-      } else {
-        _remainingChanges.push(change);
-      }
-    }
-
-    remainingChanges = _remainingChanges;
-
-    if (nextChanges.length === 0) {
-      updatedElements.push(item);
+    // When there are no changes for an element we can just push it unmodified,
+    // no need to copy it.
+    if (!changes) {
+      updatedElements.push(element);
       continue;
     }
 
-    const updateItem = { ...item };
-    let isDeletion = false;
-
-    for (const currentChange of nextChanges) {
-      if (currentChange) {
-        switch (currentChange.type) {
-          case 'select': {
-            updateItem.selected = currentChange.selected;
-            break;
-          }
-          case 'position': {
-            if (typeof currentChange.position !== 'undefined') {
-              updateItem.position = currentChange.position;
-            }
-
-            if (typeof currentChange.positionAbsolute !== 'undefined') {
-              if (!updateItem.computed) {
-                updateItem.computed = {};
-              }
-              updateItem.computed.positionAbsolute = currentChange.positionAbsolute;
-            }
-
-            if (typeof currentChange.dragging !== 'undefined') {
-              updateItem.dragging = currentChange.dragging;
-            }
-
-            if (updateItem.expandParent) {
-              handleParentExpand(updatedElements, updateItem);
-            }
-            break;
-          }
-          case 'dimensions': {
-            if (typeof currentChange.dimensions !== 'undefined') {
-              if (!updateItem.computed) {
-                updateItem.computed = {};
-              }
-              updateItem.computed.width = currentChange.dimensions.width;
-              updateItem.computed.height = currentChange.dimensions.height;
-
-              // this is needed for the node resizer to work
-              if (currentChange.resizing) {
-                updateItem.width = currentChange.dimensions.width;
-                updateItem.height = currentChange.dimensions.height;
-              }
-            }
-
-            if (typeof currentChange.resizing === 'boolean') {
-              updateItem.resizing = currentChange.resizing;
-            }
-
-            if (updateItem.expandParent) {
-              handleParentExpand(updatedElements, updateItem);
-            }
-            break;
-          }
-          case 'remove': {
-            isDeletion = true;
-            continue;
-          }
-        }
-      }
+    // If we have a 'remove' change queued, it'll be the only change in the array
+    if (changes[0].type === 'remove') {
+      continue;
     }
 
-    if (!isDeletion) {
-      updatedElements.push(updateItem);
+    if (changes[0].type === 'replace') {
+      updatedElements.push({ ...changes[0].item });
+      continue;
     }
+
+    // For other types of changes, we want to start with a shallow copy of the
+    // object so React knows this element has changed. Sequential changes will
+    /// each _mutate_ this object, so there's only ever one copy.
+    const updatedElement = { ...element };
+
+    for (const change of changes) {
+      applyChange(change, updatedElement, elements);
+    }
+
+    updatedElements.push(updatedElement);
   }
 
   return updatedElements;
+}
+
+// Applies a single change to an element. This is a *mutable* update.
+function applyChange(change: any, element: any, elements: any[] = []): any {
+  switch (change.type) {
+    case 'select': {
+      element.selected = change.selected;
+      break;
+    }
+
+    case 'position': {
+      if (typeof change.position !== 'undefined') {
+        element.position = change.position;
+      }
+
+      if (typeof change.positionAbsolute !== 'undefined') {
+        element.computed ??= {};
+        element.computed.positionAbsolute = change.positionAbsolute;
+      }
+
+      if (typeof change.dragging !== 'undefined') {
+        element.dragging = change.dragging;
+      }
+
+      if (element.expandParent) {
+        handleParentExpand(elements, element);
+      }
+      break;
+    }
+
+    case 'dimensions': {
+      if (typeof change.dimensions !== 'undefined') {
+        element.computed ??= {};
+        element.computed.width = change.dimensions.width;
+        element.computed.height = change.dimensions.height;
+
+        if (change.resizing) {
+          element.width = change.dimensions.width;
+          element.height = change.dimensions.height;
+        }
+      }
+
+      if (typeof change.resizing === 'boolean') {
+        element.resizing = change.resizing;
+      }
+
+      if (element.expandParent) {
+        handleParentExpand(elements, element);
+      }
+
+      break;
+    }
+  }
 }
 
 /**
@@ -232,6 +241,62 @@ export function getSelectionChanges(
         item.selected = willBeSelected;
       }
       changes.push(createSelectionChange(item.id, willBeSelected));
+    }
+  }
+
+  return changes;
+}
+
+/**
+ * This function is used to find the changes between two sets of elements.
+ * It is used to determine which nodes or edges have been added, removed or replaced.
+ *
+ * @internal
+ * @param params.items = the next set of elements (nodes or edges)
+ * @param params.lookup = a lookup map of the current store elements
+ * @returns an array of changes
+ */
+export function getElementsDiffChanges({
+  items,
+  lookup,
+}: {
+  items: Node[] | undefined;
+  lookup: NodeLookup;
+}): NodeChange[];
+export function getElementsDiffChanges({
+  items,
+  lookup,
+}: {
+  items: Edge[] | undefined;
+  lookup: EdgeLookup;
+}): EdgeChange[];
+export function getElementsDiffChanges({
+  items = [],
+  lookup,
+}: {
+  items: any[] | undefined;
+  lookup: Map<string, any>;
+}): any[] {
+  const changes: any[] = [];
+  const itemsLookup = new Map<string, any>(items.map((item) => [item.id, item]));
+
+  for (const item of items) {
+    const storeItem = lookup.get(item.id);
+
+    if (storeItem !== undefined && storeItem !== item) {
+      changes.push({ id: item.id, item: item, type: 'replace' });
+    }
+
+    if (storeItem === undefined) {
+      changes.push({ item: item, type: 'add' });
+    }
+  }
+
+  for (const [id] of lookup) {
+    const nextNode = itemsLookup.get(id);
+
+    if (nextNode === undefined) {
+      changes.push({ id, type: 'remove' });
     }
   }